Removed Unnecessary Bounds Checking in macro-loops.
The special "for" and "until" macro-loops performed bounds-checks before each iteration to break if the list ended before reaching the final iteration.

The purpose of these macros is to compile indistinguishably to similar hand-written versions, a.k.a a "zero-cost abstraction".

A hand written loop is unlikely to check bounds this way, thus performing better. Further, if the user wishes to iterate to a node, then a loop should do just that and no more, leaving the user to deal with any inability to reach said node as they wish. Usually, the user already knows that such a node exists so the bounds check is unnecessary baggage.

In consideration of the above, the bounds checks have been removed.